+++ src/hooks/prerun/check-dependencies.ts
@@ -49,6 +49,11 @@ export const hook = async (options: any) => {
           await git({ output: true }).addConfig("user.email", email);
           uxLog(this, `Defined ${email} as git user.email` + (email === "default@hardis-group.com") ? " (temporary)" : "");
         }
+        // Manage special characters in git file / folder names
+        if (allConfigs["core.quotepath"] == null || allConfigs["core.quotepath"] == "true" ) {
+          await git({ output: true }).addConfig("core.quotepath", "false");
+          uxLog(this, `Defined "false" as git core.quotepath`);
+        }
         // Merge tool
         if (allConfigs["merge.tool"] == null) {
           await git({ output: true }).addConfig("merge.tool", "vscode");